+++ Application/Interfaces/IRepositories/ICheckinRepository.cs
@@ -10,9 +10,9 @@ namespace Application.Interfaces.IRepositories
     public interface ICheckinRepository
     {
         Task<Checkin?> GetByIdAsync(long id);
-        Task<List<Checkin>> GetByEmployeeIdAsync(int employeeId);
-        Task<List<Checkin>> GetByEmployeeAndDateAsync(int employeeId, DateTime date);
-        Task<List<Checkin>> GetByEmployeeAndMonthAsync(int employeeId, int year, int month);
+        Task<List<Checkin>> GetByEmployeeIdAsync(long employeeId);
+        Task<List<Checkin>> GetByEmployeeAndDateAsync(long employeeId, DateTime date);
+        Task<List<Checkin>> GetByEmployeeAndMonthAsync(long employeeId, int year, int month);
         Task AddAsync(Checkin checkin);
         Task UpdateAsync(Checkin checkin);
         Task DeleteAsync(Checkin checkin);
