+++ Application/Interfaces/IServices/ICheckinService.cs
@@ -10,9 +10,9 @@ namespace Application.Interfaces.IServices
     public interface ICheckinService
     {
         Task<CheckinDto?> GetCheckinByIdAsync(long id);
-        Task<List<CheckinDto>> GetCheckinsByEmployeeAsync(int employeeId);
-        Task<List<CheckinDto>> GetCheckinsByEmployeeAndDateAsync(int employeeId, DateTime date);
-        Task<List<CheckinDto>> GetCheckinsByEmployeeAndMonthAsync(int employeeId, int year, int month);
+        Task<List<CheckinDto>> GetCheckinsByEmployeeAsync(long employeeId);
+        Task<List<CheckinDto>> GetCheckinsByEmployeeAndDateAsync(long employeeId, DateTime date);
+        Task<List<CheckinDto>> GetCheckinsByEmployeeAndMonthAsync(long employeeId, int year, int month);
         Task<CheckinDto> CreateCheckinAsync(CheckinDto checkinDto);
         Task DeleteCheckinAsync(long id);
     }
