+++ Domain/Entities/Position.cs
@@ -5,7 +5,7 @@ namespace Domain.Entities;
 
 public partial class Position
 {
-    public int PositionId { get; set; }
+    public long PositionId { get; set; }
 
     public string PositionName { get; set; } = null!;
